+++ pyunicorn/timeseries/_ext/src_numerics.c
@@ -48,7 +48,7 @@ void _euclidean_distance_matrix_fast(int ntime_x, int ntime_y, int dim,
 
 
 void _supremum_distance_matrix_fast(int ntime_x, int ntime_y, int dim, 
-    double *x_embedded, double *y_embedded, float *distance)  {
+    float *x_embedded, float *y_embedded, float *distance)  {
 
     float temp_diff, diff;
